Refactor: refactor question-cards.tsx - divide question-cards.tsx and question-card.tsx - add question card type : 'default' | 'swap' | 'check' - refactor codes to use question card type
part of SWM-129
related to #45
